Fix #56, heartbeat actions refactor
