+++ src/common/heartbeat/flow.js
@@ -12,7 +12,7 @@
 
 "use strict";
 
-let sdkobj = require("../jetpack/object");
+let sdkobj = require("../../jetpack/object");
 let merge  = sdkobj.merge;
 
 let FlowBase = function () {
