+++ src/common/heartbeat/index.js
@@ -1,0 +1,3 @@
+exports.Flow = require('./flow').Flow;
+exports.phonehome = require('./phonehome');  // or do I want to do 1 down?
+exports.showHeartbeat = require('./ui').showHeartbeat;
