+++ test/z-ensure-coverage.js
@@ -2,13 +2,14 @@
 require("../src/always.js");
 require("../src/common/actions.js");
 require("../src/common/events.js");
-require("../src/common/heartbeat-flow.js");
-require("../src/common/heartbeat-phonehome.js");
-require("../src/common/heartbeat.js");
+require("../src/common/heartbeat/flow.js");
+require("../src/common/heartbeat/index.js");
+require("../src/common/heartbeat/phonehome.js");
+require("../src/common/heartbeat/ui.js");
+require("../src/common/heartbeat/upload-validate.js");
 require("../src/common/index.js");
 require("../src/common/personinfo.js");
 require("../src/common/request.js");
-require("../src/common/upload-validate.js");
 require("../src/heartbeat-by-user-first-impression/config.js");
 require("../src/heartbeat-by-user-first-impression/heartbeat-by-user-first-impression.js");
 require("../src/repairs.js");
